+++ test-data/projects/scittle-replicant-tic-tac-toe/.joyride/src/ttt_flare.cljs
@@ -5,17 +5,17 @@
 (defn project-path [path]
   (str "projects/scittle-replicant-tic-tac-toe/" path))
 
-(defn replicant-ttt []
+(defn replicant-ttt [port]
   (flare/flare!+
    {:html [:html
            [:head
-            [:script "var SCITTLE_NREPL_WEBSOCKET_PORT = 1340;
-                      var SCITTLE_NREPL_WEBSOCKET_HOST = '127.0.0.1';"]
-            [:script {:src "https://cdn.jsdelivr.net/npm/scittle@0.7.30/dist/scittle.js"
+            [:script (str "var SCITTLE_NREPL_WEBSOCKET_PORT = " port ";
+                      var SCITTLE_NREPL_WEBSOCKET_HOST = '127.0.0.1';")]
+            [:script {:src (project-path "resources/scittle/dist/scittle.js")
                       :type "application/javascript"}]
-            [:script {:src "https://cdn.jsdelivr.net/npm/scittle@0.7.30/dist/scittle.nrepl.js"
+            [:script {:src (project-path "resources/scittle/dist/scittle.nrepl.js")
                       :type "application/javascript"}]
-            [:script {:src "https://cdn.jsdelivr.net/npm/scittle@0.7.30/dist/scittle.replicant.js"
+            [:script {:src (project-path "resources/scittle/dist/scittle.replicant.js")
                       :type "application/javascript"}]
             [:script {:type "application/x-scittle"
                       :src (project-path "resources/scittle/replicant_tictactoe/ui.cljs")}]
@@ -39,5 +39,5 @@
     :title "Greetings, Professor Falken."}))
 
 (comment
-  (replicant-ttt)
+  (replicant-ttt 1340)
   :rcf)